requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
transformers
google-generativeai
python-dotenv
PyPDF2
# faiss-cpu
gtts
pydantic
python-multipart
streamlit
groq
edge-tts
SpeechRecognition
pyaudio
